s3: Configure a clone of default HTTP transport
When the S3 storage driver is configured with skipverify: true, one
would expect that the only difference would be the skipping of TLS
server verification. But setting that parameter has the unfortunate side
effect of disabling HTTP proxy support and timeouts. These are set on
http.DefaultTransport but are not the default with a zero-valued
http.Transport. Configure skipping TLS verification by mutating a clone
of http.DefaultTransport so that proxy support, timeouts and any other
configuration which may be added to the defaults in future Go versions
are retained.

Signed-off-by: Cory Snider <csnider@mirantis.com>